+++ app/src/main/java/com/vesaa/mytv/ui/screens/leanback/panel/components/PanelIptvItem.kt
@@ -40,7 +40,6 @@ import com.vesaa.mytv.data.entities.EpgProgramme
 import com.vesaa.mytv.data.entities.EpgProgramme.Companion.isLive
 import com.vesaa.mytv.data.entities.EpgProgramme.Companion.progress
 import com.vesaa.mytv.data.entities.Iptv
-import com.vesaa.mytv.BuildConfig
 import com.vesaa.mytv.ui.components.IptvLogoImage
 import com.vesaa.mytv.ui.theme.LeanbackTheme
 import com.vesaa.mytv.ui.utils.handleLeanbackKeyEvents
@@ -126,7 +125,7 @@ fun LeanbackPanelIptvItem(
                     .padding(horizontal = 6.dp, vertical = 4.dp),
                 verticalAlignment = Alignment.CenterVertically,
             ) {
-                if (BuildConfig.CHANNEL_LOGOS_ENABLED && iptv.logoUrl.isNotBlank()) {
+                if (iptv.logoUrl.isNotBlank()) {
                     IptvLogoImage(
                         logoUrl = iptv.logoUrl,
                         contentDescription = iptv.name,